+++ src/utils/mod.rs
@@ -5,6 +5,7 @@
 pub mod dist;
 pub mod graph_utils;
 pub mod heap_structs;
+pub mod input;
 pub mod k_means_utils;
 pub mod nndescent_utils;
 pub mod parallelism;